Fix misunderstanding with Q Learning names
It is stil a bit obscure if someone knows Watkin's Q Learning as we are
using that name for the version with neural networks (which although
seems almost the same thing has really different behavior and
implications). However, Sergey Levines uses that name in both lecture 7
and 8. We may change it to "Online Q Iteration" in the future.